Updated Code and moved classes to proper packages
